Fix many type mismatches in Read/Write operations.
NOTE: Many of these casts are due to using an enum rather than a fixed
size type for field storage.

+++ Sources/Plasma/FeatureLib/pfGameGUIMgr/pfGUIControlHandlers.cpp
@@ -64,7 +64,7 @@ void    pfGUICtrlProcWriteableObject::Write( pfGUICtrlProcWriteableObject *obj, 
         obj->IWrite( s );
     }
     else
-        s->WriteLE32( kNull );
+        s->WriteLE32((uint32_t)kNull);
 }
 
 pfGUICtrlProcWriteableObject *pfGUICtrlProcWriteableObject::Read( hsStream *s )
